Tornado: delete the C encoder, and remove the DARC_NO_RUST opt-out
Deletes MatchFinder.cpp (956), OptimalParsing.cpp (44) and Tornado.cpp's
encoder section (315), plus the 80 lines of CI that existed to build the C
alongside the Rust and compare.

Two things survived MatchFinder.cpp because the remaining code uses them:
PtrVal, which C_Tornado.cpp's SetDictionary takes the sizeof, and the
check_match/print_literal/print_match debug macros, which the decompressor's
output loop calls. Both moved to Tornado.cpp where the #include was. The
encoder section was cut on its own #ifndef/#endif pair rather than by line
range -- a line range is how GRZip_CheckBlockSign was taken out along with its
neighbour earlier in this port.

OptimalParsing.cpp turns out to have been referenced by nothing at all: not
#included, not in any makefile. It was dead in every build long before this.

DARC_NO_RUST is gone from compile, compile-mhs-win64 and compile-win64-c, and
DARC_RUST is now unconditional for arc. The coverage it provided is not lost:
"build twice and compare" could only ever work while the C still existed,
whereas each rust/difftest harness compares against a *pinned* revision
fetched with git archive, so byte-identity keeps being proved after the
working-tree C is deleted. The Windows "verify the codecs linked" step now
checks for darc_rs_* symbols instead of diffing against a second build.

The #ifndef DARC_RUST blocks around the *decoders* are deliberately kept.
Unarc/makefile does not define DARC_RUST, so those are its only source for
tor_decompress, mm_decompress, tta_decompress, grzip_decompress and the
encryption entry points; only the encoder-side guards were dead. Deleting them
is a separate decision about whether Unarc lives.

Verified: arc builds, round-trips at six tor presets, and -mtor:5 produces
931364 bytes -- the same as before the deletion.
