+++ compile
@@ -57,7 +57,7 @@ if [ -n "${DARC_SANITIZE:-}" ]; then
   export DARC_DEBUG_FLAGS="-g -fsanitize=${DARC_SANITIZE} -fno-omit-frame-pointer"
   echo "sanitizer build: -fsanitize=${DARC_SANITIZE}"
 fi
-# Rust codecs: ON by default. DARC_NO_RUST=1 ./compile builds the C instead.
+# Rust codecs: mandatory. The DARC_NO_RUST opt-out was removed; see below.
 #
 # This must run BEFORE any C is compiled. C_Delta.h keys off DARC_RUST, and
 # C_Delta.cpp is built by Compression/compile below -- adding the define after
@@ -73,41 +73,31 @@ fi
 # fingerprints can prove they produce identical bytes. That comparison is what
 # licenses deleting the C, so it outlives the switch being "optional".
 rust_lib=""
-# The Rust codecs are ON by default. DARC_NO_RUST=1 opts out and builds the C
-# implementations instead -- kept so CI can build both and prove they produce
-# identical archives, and so a bisect can still reach a pure-C binary.
+# The Rust codecs are mandatory. The DARC_NO_RUST opt-out is gone: it existed
+# to build the C implementations alongside and prove the two produced identical
+# archives, and that job is done -- each ported codec is pinned byte-for-byte
+# against a fixed revision of the C by its own harness in rust/difftest, which
+# compares against `git archive` of that revision rather than the working tree.
+# So the comparison survives the C being deleted, which the opt-out did not.
 command -v cargo >/dev/null 2>&1 || {
   echo "error: cargo not found in PATH -- it is required to build DArc" >&2; exit 1; }
 
-# The Rust staticlibs are ALWAYS linked, even under DARC_NO_RUST: zstd has no C
-# implementation left, so C_Zstd.o resolves zstd_stream_* from the crate either
-# way. What DARC_NO_RUST changes is the `dropin` feature and the -DDARC_RUST
-# define -- without them the other codecs compile and bind their C, which is
-# what makes the two builds comparable.
-rust_features=""
-if [ -z "${DARC_NO_RUST:-}" ]; then
-  rust_features="--features darc-codecs/dropin"
-fi
-echo "building Rust codecs and crypto${DARC_NO_RUST:+ (dropins off)}"
-( cd rust && cargo build --release -p darc-codecs -p darc-crypto $rust_features ) || {
+echo "building Rust codecs and crypto"
+( cd rust && cargo build --release -p darc-codecs -p darc-crypto --features darc-codecs/dropin ) || {
   echo "error: cargo build failed" >&2; exit 1; }
 
-if true; then
-  # rust/ is a cargo workspace, so build output lands in the WORKSPACE target
-  # directory, not each member's. Both staticlibs are linked: codecs supply the
-  # ported compression methods, darc-crypto the encryption (docrypt, PBKDF2,
-  # random) that C_Encryption.cpp forwards to under -DDARC_RUST.
-  rust_lib="$PWD/rust/target/release/libdarc_codecs.a $PWD/rust/target/release/libdarc_crypto.a"
-  for l in $rust_lib; do
-    [ -f "$l" ] || { echo "error: cargo produced no $l" >&2; exit 1; }
-  done
-  if [ -z "${DARC_NO_RUST:-}" ]; then
-    defines="$defines -DDARC_RUST"
-    # Reaches the per-codec makefiles through unix-common.mak, which appends
-    # DARC_EXTRA_DEFINES to DEFINES; make imports it from the environment.
-    export DARC_EXTRA_DEFINES="-DDARC_RUST"
-  fi
-fi
+# rust/ is a cargo workspace, so build output lands in the WORKSPACE target
+# directory, not each member's. Both staticlibs are linked: codecs supply the
+# ported compression methods, darc-crypto the encryption (docrypt, PBKDF2,
+# random) that C_Encryption.cpp forwards to under -DDARC_RUST.
+rust_lib="$PWD/rust/target/release/libdarc_codecs.a $PWD/rust/target/release/libdarc_crypto.a"
+for l in $rust_lib; do
+  [ -f "$l" ] || { echo "error: cargo produced no $l" >&2; exit 1; }
+done
+defines="$defines -DDARC_RUST"
+# Reaches the per-codec makefiles through unix-common.mak, which appends
+# DARC_EXTRA_DEFINES to DEFINES; make imports it from the environment.
+export DARC_EXTRA_DEFINES="-DDARC_RUST"
 
 cp unix-common.mak common.mak
 cd Compression
